[CAN-13] refactor: 스키마 파일 위치 변경 - schemas -> lib, 파일 이름 변경 - auth.ts -> validation.ts
