+++ src/components/auth/TextField.tsx
@@ -1,4 +1,4 @@
-import { ValidationSchemaType } from '@/schemas/auth';
+import { ValidationSchemaType } from '@/lib/validation';
 import ErrorMessage from './ErrorMessage';
 import { FieldErrors, UseFormRegister } from 'react-hook-form';
